services/chat-docs-service/requirements.txt
fastapi
uvicorn
pydantic
qdrant-client
python-dotenv
psycopg2-binary
requests
debugpy
